make biometrics screen text and icon more generic so its method-agnostic
