+++ src/components/assets/icons/BiometricsIcon/index.tsx
@@ -1,0 +1,32 @@
+import {ColorValue, View} from 'react-native';
+import Svg, {Path, Rect, SvgProps} from 'react-native-svg';
+
+interface Props {
+  size?: number;
+  color?: ColorValue;
+}
+
+const BiometricsIcon = (props: Props) => {
+  const {color, size} = props;
+  return (
+    <View style={{width: size, aspectRatio: 1}}>
+      <Svg width="100%" height="100%" viewBox="0 0 79 65" fill="none">
+        <Path
+          fill-rule="evenodd"
+          clip-rule="evenodd"
+          d="M47.0684 56.9866C47.0714 56.9794 47.0743 56.972 47.077 56.9646C49.8398 49.5964 50.1207 41.023 47.9011 31.384C47.8197 31.0302 47.4724 30.8107 47.1256 30.8939C47.1061 30.8985 47.087 30.9041 47.0684 30.9104V33.6657C48.4396 40.6335 48.4223 46.9497 47.0684 52.5815V56.9866ZM47.0684 27.7801C47.1842 27.7688 47.3019 27.7629 47.4214 27.7629C49.2527 27.7629 50.799 29.1374 51.1404 30.9871C52.0794 36.5435 52.5684 41.696 52.2202 46.1323C52.1918 46.4946 52.4566 46.8118 52.8117 46.8408C53.1669 46.8698 53.4778 46.5997 53.5062 46.2374C53.8672 41.6391 53.3577 36.3597 52.411 30.7587L52.4094 30.7494C51.9621 28.3115 49.9112 26.4467 47.4214 26.4467C47.3024 26.4467 47.1847 26.451 47.0684 26.4595V27.7801ZM47.0684 23.3574C48.976 23.2895 50.8954 23.9129 52.3695 25.1301C55.7483 27.9201 56.2024 33.1507 56.5649 37.327L56.5651 37.3298C56.5822 37.5265 56.5991 37.7209 56.6161 37.9126C57.1286 43.7013 56.5447 49.3498 55.3468 55.0735C55.2724 55.429 55.4945 55.7786 55.843 55.8545C56.1914 55.9304 56.5341 55.7038 56.6085 55.3484C57.8247 49.5375 58.4284 43.7513 57.901 37.7941L57.8802 37.5587C57.472 32.9373 56.9684 27.2343 53.1812 24.107C51.4699 22.694 49.2633 21.9733 47.0684 22.0405V23.3574ZM47.0684 19.0267C47.1753 19.0237 47.2823 19.0221 47.3894 19.0221C50.1234 19.0221 57.197 20.5324 59.4471 28.0103C60.8571 32.6965 61.2417 39.0385 60.9002 45.1998C60.5587 51.3603 59.4964 57.2525 58.0679 61.0308C57.9397 61.3699 58.1052 61.7509 58.4376 61.8817C58.77 62.0125 59.1434 61.8436 59.2716 61.5045C60.7661 57.5515 61.8426 51.5102 62.1883 45.2741C62.5339 39.0389 62.154 32.5213 60.6804 27.6241C58.1827 19.3233 50.3531 17.7059 47.3894 17.7059C47.2824 17.7059 47.1753 17.7073 47.0684 17.7101V19.0267ZM47.0684 14.7432C47.1751 14.7411 47.2822 14.7401 47.3894 14.7401C51.5435 14.7401 55.339 16.2676 58.2522 18.7916C58.5237 19.0269 58.9308 18.993 59.1614 18.716C59.392 18.4389 59.3589 18.0236 59.0873 17.7883C55.948 15.0685 51.8587 13.4239 47.3894 13.4239C47.2822 13.4239 47.1752 13.4249 47.0684 13.4267V14.7432ZM47.0684 10.2789C52.354 10.1711 57.6431 12.0017 61.5267 15.7126C65.6845 19.6854 68.0988 25.4662 68.7308 31.7732C68.767 32.1348 69.0837 32.398 69.4381 32.361C69.7926 32.324 70.0505 32.0009 70.0143 31.6394C69.358 25.0903 66.8417 18.9879 62.4087 14.7519C58.2629 10.7905 52.6533 8.85551 47.0684 8.96278V10.2789ZM47.0684 5.72702C54.4254 5.57919 61.6806 8.77019 66.6221 14.6406C66.8543 14.9163 67.2615 14.9479 67.5318 14.7111C67.8021 14.4743 67.833 14.0588 67.6009 13.7831C62.4095 7.61583 54.7946 4.26302 47.0684 4.41057V5.72702ZM47.0684 1.36646C49.4179 1.38377 51.7588 1.69792 54.0366 2.28037C64.1485 4.86595 72.5653 12.6222 75.771 22.754C77.0108 26.6723 77.5375 30.8445 77.6429 35.217C77.6517 35.5804 77.9475 35.8677 78.3037 35.8588C78.6598 35.8498 78.9415 35.548 78.9327 35.1847C78.8258 30.7521 78.293 26.4726 77.0226 22.4249C73.6934 11.8177 64.8964 3.68705 54.3133 0.993869C51.9464 0.391559 49.5131 0.0668864 47.0684 0.0498047V1.36646ZM68.671 16.4787C68.9668 16.2761 69.3675 16.3564 69.5662 16.6582C72.7456 21.4879 74.031 26.7141 74.5345 32.7572C74.5647 33.1193 74.3014 33.4379 73.9464 33.4686C73.5914 33.4994 73.2792 33.2308 73.249 32.8687C72.7562 26.9549 71.5098 21.9715 68.4951 17.392C68.2965 17.0903 68.3752 16.6814 68.671 16.4787ZM74.1033 35.5513C74.4589 35.5286 74.7651 35.8043 74.7873 36.167C75.0305 40.1354 75.0327 44.6253 74.6024 47.9432C74.5557 48.3036 74.2315 48.557 73.8783 48.5093C73.5251 48.4617 73.2767 48.1309 73.3234 47.7706C73.7381 44.5726 73.7404 40.178 73.4997 36.2491C73.4775 35.8864 73.7477 35.5739 74.1033 35.5513ZM69.8793 34.7736C70.2337 34.7364 70.5505 34.9994 70.5869 35.361C71.1995 41.4465 70.9355 48.0194 70.0106 54.3779C69.9583 54.7375 69.6303 54.9857 69.2778 54.9323C68.9254 54.879 68.6821 54.5443 68.7344 54.1848C69.6464 47.9146 69.9033 41.4533 69.3035 35.4955C69.2671 35.1339 69.5249 34.8107 69.8793 34.7736ZM60.5326 20.2421C60.7898 19.9905 61.1981 19.9992 61.4447 20.2616C64.2915 23.2902 65.2028 28.0504 65.8058 32.0451C66.485 36.5449 66.5887 40.9347 66.5887 42.9701C66.5887 43.3336 66.2999 43.6282 65.9436 43.6282C65.5873 43.6282 65.2985 43.3336 65.2985 42.9701C65.2985 40.9677 65.1958 36.6521 64.5307 32.2455C63.9709 28.5368 63.1639 23.9922 60.5135 21.1726C60.2669 20.9102 60.2755 20.4936 60.5326 20.2421ZM65.8331 46.0308C66.1883 46.0597 66.4532 46.3768 66.4249 46.7391C66.1073 50.8037 65.6878 54.933 64.8257 58.7721C64.7461 59.1264 64.4001 59.3478 64.0529 59.2667C63.7056 59.1855 63.4886 58.8325 63.5681 58.4782C64.4081 54.7376 64.822 50.6883 65.1388 46.6345C65.1671 46.2722 65.478 46.0019 65.8331 46.0308ZM55.2308 57.8229C55.5761 57.9126 55.7847 58.2708 55.6968 58.623C55.2737 60.3186 54.6218 61.8467 53.8007 63.2463C53.6177 63.5582 53.2216 63.6597 52.9159 63.473C52.6102 63.2863 52.5107 62.8822 52.6937 62.5703C53.4556 61.2717 54.0565 59.8611 54.4465 58.2983C54.5344 57.9461 54.8855 57.7332 55.2308 57.8229ZM52.5423 49.0727C52.8961 49.1154 53.149 49.4427 53.1071 49.8036C52.4979 55.056 51.1224 59.7404 48.9155 63.8033C48.7428 64.1212 48.3502 64.2361 48.0385 64.06C47.7269 63.8838 47.6143 63.4832 47.787 63.1653C49.8991 59.277 51.2327 54.7627 51.8259 49.6489C51.8678 49.288 52.1885 49.03 52.5423 49.0727Z"
+          fill={color}
+        />
+        <Path
+          fill-rule="evenodd"
+          clip-rule="evenodd"
+          d="M32.46 22.8144C32.0197 23.1182 31.7302 23.6323 31.7302 24.2154V36.2329C31.7302 36.7506 31.3174 37.1717 30.81 37.1717H29.4602C28.5453 37.1717 27.8037 37.9284 27.8037 38.8617C27.8037 39.7951 28.5453 40.5517 29.4602 40.5517H30.81C31.395 40.5517 31.9527 40.43 32.46 40.2101V22.8144ZM32.46 48.4153C32.2168 48.4293 31.9725 48.4363 31.7274 48.4363C28.4695 48.4363 25.3584 47.1952 22.967 44.9417C22.2949 44.3085 21.247 44.3509 20.6262 45.0365C20.0055 45.7221 20.047 46.7913 20.719 47.4246C23.7244 50.2566 27.6338 51.8162 31.7274 51.8162C31.9723 51.8162 32.2165 51.8106 32.46 51.7995V48.4153ZM3.50866 10.4448V17.6428C3.50866 18.5416 2.79447 19.2702 1.91351 19.2702C1.03256 19.2702 0.318359 18.5416 0.318359 17.6428V10.4448C0.318359 4.71567 4.88712 0.0546875 10.5028 0.0546875H17.5583C18.4392 0.0546875 19.1534 0.783312 19.1534 1.68206C19.1534 2.58081 18.4392 3.30943 17.5583 3.30943H10.5028C6.64627 3.30943 3.50866 6.51035 3.50866 10.4448ZM19.1534 62.5208C19.1534 61.622 18.4392 60.8934 17.5583 60.8934H10.5028C6.64627 60.8934 3.50866 57.6924 3.50866 53.758V46.56C3.50866 45.6612 2.79447 44.9326 1.91351 44.9326C1.03256 44.9326 0.318359 45.6612 0.318359 46.56V53.758C0.318359 59.4871 4.88712 64.1481 10.5028 64.1481H17.5583C18.4392 64.1481 19.1534 63.4195 19.1534 62.5208ZM18.9385 30.5371C18.0745 30.5371 17.374 29.8225 17.374 28.941V24.1215C17.374 23.24 18.0745 22.5254 18.9385 22.5254C19.8025 22.5254 20.503 23.24 20.503 24.1215V28.941C20.503 29.8225 19.8025 30.5371 18.9385 30.5371Z"
+          fill={color}
+        />
+        <Rect x="38.3027" y="0.0498047" width="2.92207" height="64.0935" rx="0.58214" fill={color} />
+      </Svg>
+    </View>
+  );
+};
+
+export default BiometricsIcon;
