+++ src/localization/translations/en.json
@@ -6,8 +6,8 @@
   "swipe_delete_button_caption": "Delete",
   "credential_offer_accepted_toast": "Credential successfully stored",
 
-  "biometrics_title": "Do you want to enable Face ID?",
-  "biometrics_subtitle": "Use biometric login en authentication with your face for faster and easier access to your account.",
+  "biometrics_title": "Do you want to enable Biometric login?",
+  "biometrics_subtitle": "Use biometric login and authentication with your face or fingerprint for faster and easier access to your account.",
   "biometrics_description": "Your biometric information won’t be collected by Sphereon and will be used locally for verification purposes only",
   "biometrics_enable_text": "Enable now",
   "biometrics_disallow": "Maybe later",
